bounds check for new etilesprite bug
there's a new? bug where minimap crashes out
i think it was fov and lights file, line 113
index out of range
happened even before title screen had disappeared and in game too
and in the next game after that- so not just bad seed?
it could somehow be etilesprite.empty which is -1 so would be out of
range?
but how would a displaychar get set to .empty?